fixed cattegory endpoint
