+++ src/components/Pomodoro/constant.js
@@ -1,9 +1,7 @@
-import { t } from "i18next";
-
 export const TIMER_SETTINGS = {
-  [t("pomodoro.timerSetting.pomodoro")]: { time: 25 * 60 },
-  [t("pomodoro.timerSetting.shortBreak")]: { time: 5 * 60 },
-  [t("pomodoro.timerSetting.longBreak")]: { time: 15 * 60 },
+  Pomodoro: { time: 25 * 60 },
+  "Short Break": { time: 5 * 60 },
+  "Long Break": { time: 15 * 60 },
 };
 
 export const formatTime = time => {
